Changed alarm.py so that DISARM always requires a pin. There is no challenge stage. There is no deactivate. From ARMED_xxxx, ARMING, TRIGGERED, etc. only option to disarm is DISARM with pin. Fixed notifier, cancelled gsm and wired_internet -notifier can send messages to my own prosody server and connects ok with gtalk but the messages it sends do not show up-. Fixed http_translate to use the new structure of house-configuration json. Added reeportOnly flag to stop_all.sh.

+++ stop_all.sh
@@ -1,17 +1,30 @@
 #!/usr/bin/env bash
 
+# Usage. Kills all processes matching variable "targets".
+#        Run with stop_all.sh -r in order to report only and not kill any process.
+
 # set -x
 
-declare targets=("check_interface.sh" "modem.sh" "sms.sh" "alarm.py" "configuration.py" "http_translate.py" "motionwrapper.py" "notifier.py" "xmpp_proxy.py" "mosquitto_sub")
+reportOnly=$1
+
+declare targets=("check_interface.sh" "modem.sh" "sms.sh" "alarm.py" "configuration.py" "http_translate.py" "motionwrapper.py" "notifier.py" "xmpp_proxy.py") # "mosquitto_sub")
 for t in "${targets[@]}"; do
-    echo "Searching $t for killing"
+    echo '------------------------------------------------'        
+    echo "Searching $t"
     target=$(echo $t | sed 's/\./\\\./g')
     # echo "target=$target"
-    for p in $(ps aux | grep "$target" | grep -v "grep" | tr -s ' ' | cut -d ' ' -f2); do 
-        echo "killing $t process $p";
+    # for p in $(ps aux | grep "$target" | grep -v "grep" | tr -s ' ' | cut -d ' ' -f2); do
+    for p in $(ps xao pid,args | grep "$target" | grep -v "grep" | tr -s ' ' | cut -d ' ' -f1); do    
+        if [ -z "$reportOnly" ] || [[ "$reportOnly" != "-r" ]] ; then
+            echo "killing $t process $p";
+        else
+            echo "reporting $t process $p";
+        fi
         # ps aux | grep "$p" | grep -v "grep"
         ps  xao pid,ppid,pgid,sid,comm,args | grep "$p" | grep -v "grep"
         # pkill -P "$p";  #killing children of p
-        kill -TERM "$p";
+        if [ -z "$reportOnly" ] || [[ "$reportOnly" != "-r" ]] ; then
+            kill -TERM "$p";
+        fi
     done
 done
